refactor(core)!: move subsystem methods off OpenRedaction facade
detectDocument/detectDocumentFile move to @openredaction/core/documents
as functions taking IDetector; detectBatch/detectDocumentsBatch move to
@openredaction/core/workers; generateReport is replaced by
createReportGenerator().generate (the generator never used the
detector — its param is now optional); healthCheck/quickHealthCheck are
covered by createHealthChecker on @openredaction/core/health.
explain() and exportConfig() stay on the class.

detector.ts no longer imports document, reports, health, or worker
modules, so root-entry bundles exclude them.

Fixes WorkerPool worker resolution: __dirname crashed under ESM and the
probed filename (worker.js) was never emitted by the build (worker.cjs)
— worker pools were unusable from published builds. Now resolved via
import.meta.url with layout candidates, verified end-to-end from dist.

BREAKING CHANGE: OpenRedaction loses detectDocument, detectDocumentFile,
generateReport, healthCheck, quickHealthCheck, and the static batch
methods; see changeset migration table.
